+++ core/jvm/src/test/scala/io/github/scala_tessella/research_core/HoneycombGateSmokeSpec.scala
@@ -37,6 +37,8 @@ class HoneycombGateSmokeSpec extends AnyFlatSpec with Matchers:
     acc.patterns.foreach { (_, pat) =>
       val cert = CompletenessAudit.debugCertify(acc, pat).get
       cert.generatorEquivariant shouldBe true
+      cert.boxPeriodic shouldBe true
+      cert.closed shouldBe true
       cert.ok shouldBe true
     }
     flags.items shouldBe empty